Modified 2-str_concat.c

+++ 0x0B-malloc_free/2-str_concat.c
@@ -14,8 +14,8 @@ char *str_concat(char *s1, char *s2)
 	if (!s2)
 		s2 = "";
 
-	unsigned int len1 = strlen(s1);
-	unsigned int len2 = strlen(s2);
+	size_t len1 = strlen(s1);
+	size_t len2 = strlen(s2);
 	char *result;
 
 	result = (char *) malloc(len1 + len2 + 1);
